projects/customer-service/requirements.txt
# 智能客服系统 - Python 依赖
# 安装：pip install -r requirements.txt

# LLM API
openai>=1.0.0

# 向量数据库
faiss-cpu>=1.7.4

# Embedding 模型
sentence-transformers>=2.2.2

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 监控
prometheus-client>=0.19.0

# 日志
python-json-logger>=2.0.0
